+++ Engine/Fleur/Renderer/Skybox.h
@@ -31,7 +31,7 @@ public:
 
     uint32_t GetVertexCount() const
     {
-        return m_Vertices.size();
+        return static_cast<uint32_t>(m_Vertices.size());
     };
 
     const float* Data() const
