+++ Engine/Fleur/Renderer/VertexLayout.h
@@ -78,7 +78,7 @@ public:
 
     uint32_t GetSizeOfDataType(EDataType type) const;
 
-    inline uint32_t Stride() const
+    inline uint16_t Stride() const
     {
         return m_Stride;
     }
@@ -91,7 +91,7 @@ public:
 
 private:
     std::vector<VertexAttribute> m_Attribs;
-    uint32_t m_Stride;
+    uint16_t m_Stride;
     LayoutIterator* m_It;
 };
 }  // namespace Fleur::Graphics
